requirements.txt
pillow>=9.5.0
matplotlib>=3.7.0
pandas>=2.0.0
openpyxl>=3.1.0